+++ src/main/php/web/Routing.class.php
@@ -138,9 +138,4 @@ class Routing implements Handler {
   public function handle($request, $response) {
     return $this->route($request)->handle($request, $response);
   }
-
-  /** @deprecated */
-  public function service($request, $response) {
-    return $this->route($request)->handle($request, $response);
-  }
 }