+++ src/main/php/web/routing/Target.class.php
@@ -4,7 +4,7 @@
  * Matches a request method and, optionally a target
  *
  * @deprecated
- * @test  xp://web.unittest.routing.TargetTest
+ * @test  web.unittest.routing.TargetTest
  */
 class Target implements RouteMatch {
   private $methods, $target;